Unhide read more section and refactor description text to separate component
